Replace videos in playtest-20221203 news post with optimized versions
The videos were resized to a maximum of 1200px in width (this is double
the size at which they are displayed to account for high-res displays).

The resizing naturally reduced the filesizes of the videos. But the biggest
gains come from tweaking the codec settings.

OGV is removed in favor of WebM - nowadays WebM + MPEG4 should cover 99% of
browser/OS combinations.